requirements.txt
openai
google-generativeai
pandas
pyyaml
requests
tqdm
tenacity
colorama
openpyxl
tiktoken
ollama
torch
transformers
accelerate
bitsandbytes
filelock